+++ .github/workflows/build-packages-web-forms.yml
@@ -8,18 +8,20 @@ on:
       - develop
     paths:
       - '.github/workflows/build-packages-web-forms.yml'
-      - 'packages/web-forms/**'
-      - 'packages/eslint-config-custom/**'
-      - 'packages/tsconfig/**'
+      - 'packages/web/forms/**'
+      - 'packages/web/ui/**'
+      - 'packages/config/eslint/**'
+      - 'packages/config/tsconfig/**'
   pull_request:
     branches:
       - main
       - develop
     paths:
       - '.github/workflows/build-packages-web-forms.yml'
-      - 'packages/web-forms/**'
-      - 'packages/eslint-config-custom/**'
-      - 'packages/tsconfig/**'
+      - 'packages/web/forms/**'
+      - 'packages/web/ui/**'
+      - 'packages/config/eslint/**'
+      - 'packages/config/tsconfig/**'
 
 jobs:
   build:
@@ -41,7 +43,7 @@ jobs:
           version: 7.9.5
 
       - name: Install dependencies
-        run: pnpm --filter web-forms... install
+        run: pnpm --filter @acme/web-forms... install
 
       - name: Lint
-        run: pnpm --filter web-forms lint
+        run: pnpm --filter @acme/web-forms lint
