+++ .github/workflows/build-packages-web-storage.yml
@@ -8,18 +8,18 @@ on:
       - develop
     paths:
       - '.github/workflows/build-packages-web-storage.yml'
-      - 'packages/web-storage/**'
-      - 'packages/eslint-config-custom/**'
-      - 'packages/tsconfig/**'
+      - 'packages/web/storage/**'
+      - 'packages/config/eslint/**'
+      - 'packages/config/tsconfig/**'
   pull_request:
     branches:
       - main
       - develop
     paths:
       - '.github/workflows/build-packages-web-storage.yml'
-      - 'packages/web-storage/**'
-      - 'packages/eslint-config-custom/**'
-      - 'packages/tsconfig/**'
+      - 'packages/web/storage/**'
+      - 'packages/config/eslint/**'
+      - 'packages/config/tsconfig/**'
 
 jobs:
   build:
@@ -41,7 +41,7 @@ jobs:
           version: 7.9.5
 
       - name: Install dependencies
-        run: pnpm --filter web-storage... install
+        run: pnpm --filter @acme/web-storage... install
 
       - name: Lint
-        run: pnpm --filter web-storage lint
+        run: pnpm --filter @acme/web-storage lint
