+++ apps/web/package.json
@@ -11,6 +11,7 @@
     "test": "jest"
   },
   "dependencies": {
+    "@acme/web-forms": "workspace:*",
     "@acme/web-storage": "workspace:*",
     "@hookform/resolvers": "^2.9.10",
     "@tanstack/react-query": "^4.3.9",
@@ -29,7 +30,6 @@
     "superjson": "^1.10.0",
     "theme": "workspace:*",
     "trpc": "workspace:*",
-    "web-forms": "workspace:*",
     "web-locales": "workspace:*",
     "web-ui": "workspace:*"
   },
