+++ packages/web/forms/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "web-forms",
+  "name": "@acme/web-forms",
   "version": "0.0.0",
   "license": "MIT",
   "scripts": {
